Start refactor of Spdy request/response into base entity
